bootstrap "target" support
- with DI, we can now "inject" primitives and services into different parts of the app
- this allows us to target multiple environments - e.g. node, CloudFlare Workers, Deno, etc.
- there's still some tight coupling with node.js which will be uncoupled over time
