+++ src/index.ts
@@ -1,14 +1,1 @@
-import express from 'express';
-import logging from '@tryghost/logging';
-import {config} from './services/config.js';
-import {logRequest} from './controllers/middleware.js';
-import {routingManager} from './routing.js';
-
-export const app = express();
-
-app.use(logRequest);
-routingManager.addAllRoutes(app);
-
-app.listen(config.port, config.hostname, () => {
-	logging.info(`Listening on http://${config.hostname}:${config.port}`);
-});
+import './targets/node.js'; // eslint-disable-line import/no-unassigned-import
