+++ src/targets/node.ts
@@ -1,0 +1,52 @@
+import {Buffer} from 'node:buffer';
+import {randomFillSync} from 'node:crypto';
+import process from 'node:process';
+import {config as loadEnv} from 'dotenv';
+import express from 'express';
+import logging from '@tryghost/logging';
+import {Configuration} from '../types/config.js';
+import {getConfig} from '../services/config.js';
+import {logRequest} from '../controllers/middleware.js';
+import {getRoutingManager} from '../services/dependency-injection.js';
+
+loadEnv();
+
+const getRandomHex = () => randomFillSync(Buffer.alloc(12)).toString('hex');
+
+const envToConfigMapping: Record<keyof Configuration, string> = {
+	hostname: 'DOG_HOSTNAME',
+	port: 'DOG_PORT',
+	discourseSecret: 'DOG_DISCOURSE_SHARED_SECRET',
+	discourseUrl: 'DOG_DISCOURSE_URL',
+	discourseApiKey: 'DOG_DISCOURSE_API_KEY',
+	discourseApiUser: 'DOG_DISCOURSE_API_USER',
+	ghostUrl: 'DOG_GHOST_URL',
+	mountedPublicUrl: 'DOG_GHOST_URL',
+	mountedBasePath: 'DOG_GHOST_URL',
+	ghostApiKey: 'DOG_GHOST_ADMIN_TOKEN',
+	logDiscourseRequests: 'DOG_LOG_DISCOURSE_REQUESTS',
+	logGhostRequests: 'DOG_LOG_GHOST_REQUESTS',
+	enableGhostWebhooks: 'DOG_GHOST_MEMBER_WEBHOOKS_ENABLED',
+	ghostMemberUpdatedRoute: 'DOG_GHOST_MEMBER_UPDATED_WEBHOOK_ID',
+	ghostMemberDeletedRoute: 'DOG_GHOST_MEMBER_DELETED_WEBHOOK_ID',
+	ghostMemberDeleteDiscourseAction: 'DOG_GHOST_MEMBER_DELETE_DISCOURSE_ACTION',
+	ssoMethod: 'DOG_DISCOURSE_SSO_TYPE',
+	noAuthRedirect: 'DOG_SSO_NO_AUTH_REDIRECT',
+} as const;
+
+const config = getConfig(process.env, envToConfigMapping, getRandomHex);
+
+if (!config) {
+	process.exit(1); // eslint-disable-line unicorn/no-process-exit
+}
+
+const routingManager = getRoutingManager(config);
+
+export const app = express();
+
+app.use(logRequest);
+routingManager.addAllRoutes(app);
+
+app.listen(config.port, config.hostname, () => {
+	logging.info(`Listening on http://${config.hostname}:${config.port}`);
+});
